# backend/docs/swagger.yaml
basePath: /
definitions:
...
  correctiveaction.createBody:
    properties:
      actionType:
        type: string
      assigneeId:
        type: string
      description:
        type: string
      dueDate:
        type: string
      priority:
        type: string
    type: object
  correctiveaction.patchBody:
    properties:
      actionType:
        type: string
      assigneeId:
        type: string
      completionNotes:
        type: string
      description:
        type: string
      dueDate:
        type: string
      priority:
        type: string
      status:
        type: string
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      success:
        type: boolean
    type: object
  response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: K3 incident reporting and compliance API (MVP).
  title: Aegis API
  version: "1.0"
paths:
...
  /corrective-actions:
    get:
      parameters:
      - description: status
        in: query
        name: status
        type: string
      - description: assignee
        in: query
        name: assigneeId
        type: string
      - description: priority
        in: query
        name: priority
        type: string
      - description: location
        in: query
        name: locationId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: List corrective actions across incidents
      tags:
      - corrective-actions
  /corrective-actions/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: ca id
        in: path
        name: id
        required: true
        type: string
      - description: patch
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/correctiveaction.patchBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Update corrective action
      tags:
      - corrective-actions
  /corrective-actions/{id}/verify:
    post:
      parameters:
      - description: ca id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Verify completed corrective action
      tags:
      - corrective-actions
  /dashboard/heatmap:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Heatmap site x category
      tags:
      - dashboard
  /dashboard/summary:
    get:
      parameters:
      - description: RFC3339
        in: query
        name: from
        type: string
      - description: RFC3339
        in: query
        name: to
        type: string
      - description: location id
        in: query
        name: locationId
        type: string
      - description: category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Dashboard KPI summary
      tags:
      - dashboard
  /dashboard/trends:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Incident trends 12 months
      tags:
      - dashboard
...
  /incidents/{id}/corrective-actions:
    get:
      parameters:
      - description: incident id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: List corrective actions for an incident
      tags:
      - corrective-actions
    post:
      consumes:
      - application/json
      parameters:
      - description: incident id
        in: path
        name: id
        required: true
        type: string
      - description: ca
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/correctiveaction.createBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Create corrective action
      tags:
      - corrective-actions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"